fix: remove components with old logic
