+++ src/modules/Admin/index.ts
@@ -1,6 +1,4 @@
 import AdminProducts from './ui/Products'
 import ProductEdit from './ui/ProductEdit/ProductEdit'
-import AdminProductsTest from './ui/Products/ProductsTest'
-import ProductTestEdit from './ui/ProductEdit/ProductTestEdit'
 
-export { AdminProducts, ProductEdit, AdminProductsTest, ProductTestEdit }
+export { AdminProducts, ProductEdit }
