+++ src/modules/Designs/index.tsx
@@ -1,4 +1,3 @@
 import DesignsContent from './ui/DesignsContent'
-import DesignsContentTest from './ui/DesignsContentTest'
 
-export { DesignsContent, DesignsContentTest }
+export { DesignsContent }
