+++ src/modules/Designs/ui/DesignsContent.tsx
@@ -1,20 +1,22 @@
 import { FC, useCallback, useState } from 'react'
-import { CategoryProps, ProductProps } from '@/utils/types'
-import HeroInner from '@/ui/HeroInner/HeroInner'
+import { CategoryProps, ProductTestProps } from '@/utils/types'
+
 import Container from '@/app/layouts/Container'
-import ProductList from '@/modules/Designs/ui/ProductList/ProductList'
-import ProductFilter from '@/modules/Designs/ui/ProductFilter/ProductFilter'
-import ProductSearch from '@/modules/Designs/ui/ProductSearch/ProductSearch'
 import BecomeDistributorSection from '@/components/BecomeDistributorSection/BecomeDistributorSection'
+import HeroInner from '@/ui/HeroInner/HeroInner'
+
+import ProductList from './ProductList/ProductList'
+import ProductFilter from './ProductFilter/ProductFilter'
+import ProductSearch from './ProductSearch/ProductSearch'
 
 import styles from './DesignsContent.module.scss'
 
-type DesignsProps = {
+type DesignsContentProps = {
   categories: CategoryProps[]
-  products: ProductProps[]
+  products: ProductTestProps[]
 }
 
-const DesignsContent: FC<DesignsProps> = ({ categories, products }) => {
+const DesignsContent: FC<DesignsContentProps> = ({ categories, products }) => {
   const [filteredCategories, setFilteredCategories] =
     useState<CategoryProps[]>(categories)
   const [searchQuery, setSearchQuery] = useState<string>('')
@@ -27,8 +29,11 @@ const DesignsContent: FC<DesignsProps> = ({ categories, products }) => {
     setSearchQuery(query)
   }, [])
 
-  const filterProductsBySearch = (products: ProductProps[], query: string) => {
-    return products.filter((product: ProductProps) =>
+  const filterProductsBySearch = (
+    products: ProductTestProps[],
+    query: string
+  ) => {
+    return products.filter((product: ProductTestProps) =>
       product.ProductName.toLowerCase().includes(query.toLowerCase())
     )
   }
