+++ src/modules/DesignsDetail/index.tsx
@@ -1,4 +1,3 @@
-import DesignsDetailContentTest from './ui/DesignDetailContentTest'
 import DesignsDetailContent from './ui/DesignsDetailContent'
 
-export { DesignsDetailContent, DesignsDetailContentTest }
+export { DesignsDetailContent }
