+++ src/modules/DesignsDetail/ui/DesignsDetailContent.tsx
@@ -3,18 +3,19 @@ import dynamic from 'next/dynamic'
 import { useRouter } from 'next/router'
 import Link from 'next/link'
 import axios from 'axios'
-import { ProductProps } from '@/utils/types'
+import { motion } from 'framer-motion'
+
+import routes from '@/utils/routes'
+import { ProductTestProps } from '@/utils/types'
+
 import Container from '@/app/layouts/Container'
 import { ButtonPrimary } from '@/ui/ButtonPrimary/ButtonPrimary'
-import routes from '@/utils/routes'
-import DesignsDetailDescription from '@/modules/DesignsDetail/ui/DesignsDetailDescription/DesignsDetailDescription'
-import DesignsDetailThumbnail from '@/modules/DesignsDetail/ui/DesignsDetailThumbnail/DesignsDetailThumbnail'
-import { motion } from 'framer-motion'
+
+import DesignsDetailDescription from './DesignsDetailDescription/DesignsDetailDescription'
+import DesignsDetailThumbnail from './DesignsDetailThumbnail/DesignsDetailThumbnail'
+
 const DesignsDetailGallery = dynamic(
-  () =>
-    import(
-      '@/modules/DesignsDetail/ui/DesignsDetailGallery/DesignsDetailGallery'
-    ),
+  () => import('./DesignsDetailGallery/DesignsDetailGallery'),
   { ssr: false }
 )
 const BecomeDistributorSection = dynamic(
@@ -33,13 +34,13 @@ const DesignsDetailContent = () => {
   const {
     query: { productId },
   } = useRouter()
-  const [product, setProduct] = useState<ProductProps | null>(null)
-  const [sameProducts, setSameProducts] = useState<ProductProps[]>([])
+  const [product, setProduct] = useState<ProductTestProps | null>(null)
+  const [sameProducts, setSameProducts] = useState<ProductTestProps[]>([])
 
   useEffect(() => {
     const fetchProduct = async () => {
       try {
-        const res = await axios.get(`/api/products/${productId}`)
+        const res = await axios.get(`/api/producttest/${productId}`)
         setProduct(res.data.data)
       } catch (error) {
         console.log(error)
@@ -54,10 +55,10 @@ const DesignsDetailContent = () => {
       if (product) {
         try {
           const res = await axios.get(
-            `/api/products?category=${product?.category?.id}`
+            `/api/producttest?category=${product?.category?.id}`
           )
           setSameProducts(
-            res.data.data.filter((p: ProductProps) => p.id !== productId)
+            res.data.data.filter((p: ProductTestProps) => p.id !== productId)
           )
         } catch (error) {
           console.log(error)
