+++ src/ui/ProductCarousel/ProductCarousel.tsx
@@ -2,14 +2,15 @@ import { FC, Fragment, useEffect, useState } from 'react'
 import dynamic from 'next/dynamic'
 import classNames from 'classnames'
 import { Splide, SplideSlide, SplideTrack } from '@splidejs/react-splide'
-import { ProductProps } from '@/utils/types'
+import { ProductTestProps } from '@/utils/types'
+
 import ProductCard from '@/ui/ProductCard/ProductCard'
 const AnimatedText = dynamic(() => import('@/ui/AnimatedText/AnimatedText'))
 
 import styles from './ProductCarousel.module.scss'
 
 type ProductCarouselProps = {
-  data?: ProductProps[]
+  data?: ProductTestProps[]
   className?: string
   title?: string
   titleTag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
